+++ Application/ResearchDataManagementPlatform/Menus/RDMPTopMenuStripUI.cs
@@ -15,6 +15,7 @@ using Rdmp.Core.CommandExecution;
 using Rdmp.Core.CommandExecution.AtomicCommands;
 using Rdmp.Core.CommandExecution.AtomicCommands.CatalogueCreationCommands;
 using Rdmp.Core.CommandExecution.AtomicCommands.CohortCreationCommands;
+using Rdmp.Core.CommandExecution.AtomicCommands.DataViewing;
 using Rdmp.Core.CommandLine.Options;
 using Rdmp.Core.Curation.Data.Cohort;
 using Rdmp.Core.Databases;