+++ Rdmp.Core/CommandExecution/AtomicCommands/DataViewing/ExecuteCommandViewCohortIdentificationConfiguration.cs
@@ -4,20 +4,34 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using Rdmp.Core.Curation.Data;
 using Rdmp.Core.Curation.Data.Cohort;
 using Rdmp.Core.DataViewing;
 using Rdmp.Core.Icons.IconProvision;
+using Rdmp.Core.Repositories.Construction;
 using ReusableLibraryCode.Icons.IconProvision;
 using System.Drawing;
+using System.IO;
 
-namespace Rdmp.Core.CommandExecution.AtomicCommands
+namespace Rdmp.Core.CommandExecution.AtomicCommands.DataViewing
 {
-    public class ExecuteCommandViewCohortIdentificationConfiguration : BasicCommandExecution, IAtomicCommand
+    /// <summary>
+    /// Run the SQL for a full <see cref="CohortIdentificationConfiguration"/> and which should
+    /// return a list of identifiers from live datasets/query cache
+    /// </summary>
+    public class ExecuteCommandViewCohortIdentificationConfiguration : ExecuteCommandViewDataBase, IAtomicCommand
     {
         private readonly CohortIdentificationConfiguration _cic;
         private readonly bool _useCache;
 
-        public ExecuteCommandViewCohortIdentificationConfiguration(IBasicActivateItems activator, CohortIdentificationConfiguration cic, bool useCache = true) : base(activator)
+        [UseWithObjectConstructor]
+        public ExecuteCommandViewCohortIdentificationConfiguration(IBasicActivateItems activator,
+            [DemandsInitialization("The cohort builder query you wan to execute")]
+            CohortIdentificationConfiguration cic,
+            [DemandsInitialization("True (default) to use the query cache (if available) or false to run suppressing caching")]
+            bool useCache = true,
+            [DemandsInitialization(ToFileDescription)]
+            FileInfo toFile = null) : base(activator,toFile)
         {
             this._cic = cic;
             this._useCache = useCache;
@@ -35,15 +49,12 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands
         {
             return iconProvider.GetImage(RDMPConcept.SQL);
         }
-        public override void Execute()
+        protected override IViewSQLAndResultsCollection GetCollection()
         {
-            base.Execute();
-
-            BasicActivator.ShowData(
-                new ViewCohortIdentificationConfigurationSqlCollection(_cic)
-                {
-                    UseQueryCache = _useCache
-                });
+            return new ViewCohortIdentificationConfigurationSqlCollection(_cic)
+            {
+                UseQueryCache = _useCache
+            };
         }
     }
 }