+++ Rdmp.Core/CommandExecution/AtomicCommands/DataViewing/ExecuteCommandViewCohortSample.cs
@@ -6,74 +6,45 @@
 
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.DataExport.Data;
-using Rdmp.Core.DataExport.DataExtraction;
 using Rdmp.Core.DataViewing;
 using System.IO;
 
-namespace Rdmp.Core.CommandExecution.AtomicCommands
+namespace Rdmp.Core.CommandExecution.AtomicCommands.DataViewing
 {
     /// <summary>
     /// Fetches the private and release identifiers for a given <see cref="ExtractableCohort"/> optionally to file
     /// </summary>
-    class ExecuteCommandViewCohortSample : BasicCommandExecution
+    class ExecuteCommandViewCohortSample : ExecuteCommandViewDataBase
     {
         private readonly bool _includeCohortID;
 
         public ExtractableCohort Cohort { get; }
         public int Sample { get; }
-        public FileInfo ToFile { get; }
 
-        /// <summary>
-        /// Set to true to prompt user to pick a <see cref="ToFile"/> at command execution time.
-        /// </summary>
-        public bool AskForFile { get; set; }
 
         public ExecuteCommandViewCohortSample(IBasicActivateItems activator,
             [DemandsInitialization("The cohort that you want to fetch records for")]
             ExtractableCohort cohort,
             [DemandsInitialization("Optional. The maximum number of records to retrieve")]
             int sample = 100,
-            [DemandsInitialization("Optional. A file to write the records to instead of the console")]
+            [DemandsInitialization(ToFileDescription)]
             FileInfo toFile = null,
 
             [DemandsInitialization("True to include the OriginId of the cohort when extracting",DefaultValue = true)]
-            bool includeCohortID = true):base(activator)
+            bool includeCohortID = true):base(activator,toFile)
         {
             Cohort = cohort;
             Sample = sample;
-            ToFile = toFile;
             _includeCohortID = includeCohortID;
         }
-        public override void Execute()
-        {
-            base.Execute();
 
-            var collection = new ViewCohortExtractionUICollection(Cohort) {
+        protected override IViewSQLAndResultsCollection GetCollection()
+        {
+            return new ViewCohortExtractionUICollection(Cohort)
+            {
                 Top = Sample,
                 IncludeCohortID = _includeCohortID
             };
-
-            FileInfo toFile = ToFile;
-
-            if(AskForFile)
-            {
-                toFile = BasicActivator.SelectFile("Save cohort as", "Comma Separated Values", "*.csv");
-                if(toFile == null)
-                {
-                    // user cancelled selecting a file
-                    return;
-                }
-            }
-
-            if(toFile == null)
-            {
-                BasicActivator.ShowData(collection);
-            }
-            else
-            {
-                ExtractTableVerbatim.ExtractDataToFile(collection, toFile);
-            }
-            
         }
     }
 }